+++ CHANGELOG.md
@@ -9,6 +9,20 @@ the "Changelog Release" GitHub Actions workflow.
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-08-25
+
+### Added
+
+- Selects: Added support for `corn-select` styles and documentation examples.
+
+
+## [0.2.0] - 2026-08-04
+
+### Added
+
+- Popovers: Added support for additional popover positions: `top-left`, `top-right`, `right-top`, `right-bottom`, `bottom-left`, `bottom-right`, `left-top`, and `left-bottom`.
+
+
 ## [0.1.1] - 2026-06-16
 
 ### Fixed